function type int to double

+++ window_scene/intention_event/src/intention_event_manager.cpp
@@ -50,7 +50,7 @@ void LogPointInfo(const std::shared_ptr<MMI::PointerEvent>& pointerEvent)
         TLOGD(WmsLogTag::WMS_EVENT, "action point info:windowid:%{public}d,id:%{public}d,displayx:%{public}d,"
             "displayy:%{public}d, windowx:%{public}d, windowy :%{public}d, action :%{public}d pressure: "
             "%{public}f, tiltx :%{public}f, tiltY :%{public}f",
-            windowId, actionId, item.GetDisplayX(), item.GetDisplayY(), item.GetWindowX(), item.GetWindowY(),
+            windowId, actionId, static_cast<int32_t>(item.GetDisplayX()), static_cast<int32_t>(item.GetDisplayY()), static_cast<int32_t>(item.GetWindowX()), static_cast<int32_t>(item.GetWindowY()),
             pointerEvent->GetPointerAction(), item.GetPressure(), item.GetTiltX(), item.GetTiltY());
     }
     auto ids = pointerEvent->GetPointerIds();
@@ -59,7 +59,7 @@ void LogPointInfo(const std::shared_ptr<MMI::PointerEvent>& pointerEvent)
         if (pointerEvent->GetPointerItem(id, item)) {
             TLOGD(WmsLogTag::WMS_EVENT, "all point info: id: %{public}d, x:%{public}d, y:%{public}d, "
                 "isPressend:%{public}d, pressure:%{public}f, tiltX:%{public}f, tiltY:%{public}f",
-            actionId, item.GetWindowX(), item.GetWindowY(), item.IsPressed(), item.GetPressure(),
+            actionId, static_cast<int32_t>(item.GetWindowX()), static_cast<int32_t>(item.GetWindowY()), item.IsPressed(), item.GetPressure(),
             item.GetTiltX(), item.GetTiltY());
         }
     }
